moved Filtering/Filter to connectors.filtering
