+++ connectors/byoc.py
@@ -8,12 +8,12 @@ Implementation of BYOC protocol.
 """
 import socket
 from collections import UserDict
-from copy import deepcopy
 from datetime import datetime, timezone
 from enum import Enum
 
 from connectors.es import ESDocument, ESIndex
 from connectors.es.client import with_concurrency_control
+from connectors.filtering import Filter, Filtering
 from connectors.filtering.validation import (
     FilteringValidationState,
     InvalidFilteringError,
@@ -286,70 +286,6 @@ class SyncJob(ESDocument):
         if len(connector_metadata) > 0:
             doc["metadata"] = connector_metadata
         await self.index.update(doc_id=self.id, doc=doc)
-
-
-class Filtering:
-    DEFAULT_DOMAIN = "DEFAULT"
-
-    def __init__(self, filtering=None):
-        if filtering is None:
-            filtering = []
-
-        self.filtering = filtering
-
-    def get_active_filter(self, domain=DEFAULT_DOMAIN):
-        return self.get_filter(filter_state="active", domain=domain)
-
-    def get_draft_filter(self, domain=DEFAULT_DOMAIN):
-        return self.get_filter(filter_state="draft", domain=domain)
-
-    def get_filter(self, filter_state="active", domain=DEFAULT_DOMAIN):
-        return next(
-            (
-                Filter(filter_[filter_state])
-                for filter_ in self.filtering
-                if filter_["domain"] == domain
-            ),
-            Filter(),
-        )
-
-    def to_list(self):
-        return list(self.filtering)
-
-
-class Filter(dict):
-    def __init__(self, filter_=None):
-        if filter_ is None:
-            filter_ = {}
-
-        super().__init__(filter_)
-
-        self.advanced_rules = filter_.get("advanced_snippet", {})
-        self.basic_rules = filter_.get("rules", [])
-        self.validation = filter_.get(
-            "validation", {"state": FilteringValidationState.VALID.value, "errors": []}
-        )
-
-    def get_advanced_rules(self):
-        return self.advanced_rules.get("value", {})
-
-    def has_advanced_rules(self):
-        advanced_rules = self.get_advanced_rules()
-        return advanced_rules is not None and len(advanced_rules) > 0
-
-    def has_validation_state(self, validation_state):
-        return FilteringValidationState(self.validation["state"]) == validation_state
-
-    def transform_filtering(self):
-        """
-        Transform the filtering in .elastic-connectors to filtering ready-to-use in .elastic-connectors-sync-jobs
-        """
-        # deepcopy to not change the reference resulting in changing .elastic-connectors filtering
-        filtering = (
-            {"advanced_snippet": {}, "rules": []} if len(self) == 0 else deepcopy(self)
-        )
-
-        return filtering
 
 
 PIPELINE_DEFAULT = {
